+++ packages/terra-button-group/src/terra-dev-site/test/button-group/ButtonGroupIcon.test.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+d import React from 'react';
 import ButtonGroup from '../../../ButtonGroup';
 
 const icon = <svg width="1em" height="1em"><rect width="1em" height="1em" fill="#000" /></svg>;