+++ packages/terra-dialog/tests/jest/Dialog.test.jsx
@@ -23,13 +23,3 @@ describe('Dialog', () => {
     expect(wrapper).toMatchSnapshot();
   });
 });
-
-it('throws error on missing locale prop in Base', () => {
-  global.console = { error: jest.fn() };
-
-  expect(() => {
-    render(<Dialog header="Header Content" footer="Footer Content" />);
-  }).toThrowError();
-  // eslint-disable-next-line no-console
-  expect(console.error).toBeCalledWith(expect.stringContaining('Component is internationalized, and must be wrapped in terra-base'));
-});