+++ packages/terra-form-textarea/tests/wdio/form-textarea-spec.js
@@ -80,8 +80,6 @@ Terra.describeViewports('Form-Textarea', ['tiny', 'large'], () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea');
         browser.moveToObject('#root', 0, 0);
-        // Removes the blinking cursor to prevent screenshot mismatches.
-        browser.execute('document.querySelector("textarea").style.caretColor = "transparent";');
       });
 
       Terra.it.validatesElement('0', { rules: ignoredA11y });
@@ -96,8 +94,6 @@ Terra.describeViewports('Form-Textarea', ['tiny', 'large'], () => {
     describe('Does not resize textareas that are non-resizable', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/medium-textarea');
-        // Removes the blinking cursor to prevent screenshot mismatches.
-        browser.execute('document.querySelector("textarea").style.caretColor = "transparent";');
       });
 
       Terra.it.validatesElement('0', { rules: ignoredA11y });
@@ -113,8 +109,6 @@ Terra.describeViewports('Form-Textarea', ['tiny', 'large'], () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea-mobile-on-touch-start');
         browser.moveToObject('#root', 0, 0);
-        // Removes the blinking cursor to prevent screenshot mismatches.
-        browser.execute('document.querySelector("textarea").style.caretColor = "transparent";');
       });
 
       Terra.it.validatesElement('0', { rules: ignoredA11y });
@@ -129,9 +123,8 @@ Terra.describeViewports('Form-Textarea', ['tiny', 'large'], () => {
     describe('Blocks autoresizing for when DocumentTouch exists and browser screen is not desktop size', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea-mobile-document-touch');
+
         browser.moveToObject('#root', 0, 0);
-        // Removes the blinking cursor to prevent screenshot mismatches.
-        browser.execute('document.querySelector("textarea").style.caretColor = "transparent";');
       });
 
       Terra.it.validatesElement('0', { rules: ignoredA11y });
@@ -147,8 +140,6 @@ Terra.describeViewports('Form-Textarea', ['tiny', 'large'], () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea-mobile-max-touch-points');
         browser.moveToObject('#root', 0, 0);
-        // Removes the blinking cursor to prevent screenshot mismatches.
-        browser.execute('document.querySelector("textarea").style.caretColor = "transparent";');
       });
 
       Terra.it.validatesElement('0', { rules: ignoredA11y });
@@ -164,8 +155,6 @@ Terra.describeViewports('Form-Textarea', ['tiny', 'large'], () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea-mobile-ms-max-touch-points');
         browser.moveToObject('#root', 0, 0);
-        // Removes the blinking cursor to prevent screenshot mismatches.
-        browser.execute('document.querySelector("textarea").style.caretColor = "transparent";');
       });
 
       Terra.it.validatesElement('0', { rules: ignoredA11y });
@@ -175,6 +164,17 @@ Terra.describeViewports('Form-Textarea', ['tiny', 'large'], () => {
       });
 
       Terra.it.matchesScreenshot('1');
+    });
+
+    describe('Swapping two textareas with different size', () => {
+      before(() => {
+        browser.url('/#/raw/tests/terra-form-textarea/form-textarea/swap-textarea');
+        // Removes the blinking cursor to prevent screenshot mismatches.
+        browser.execute('document.querySelector("textarea").style.caretColor = "transparent";');
+        browser.click('#swap-button');
+      });
+
+      Terra.it.validatesElement({ rules: ignoredA11y });
     });
   });