+++ packages/terra-overlay/src/LoadingOverlay.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import IconSpinner from 'terra-icon/lib/icon/IconSpinner';
-
+import { FormattedMessage } from 'react-intl';
 import Overlay from './Overlay';
 import styles from './Overlay.module.scss';
 
@@ -12,25 +12,25 @@ const { BackgroundStyles, zIndexes } = Overlay.Opts;
 
 const propTypes = {
   /**
- * The message to be displayed within the overlay.
- */
-  message: PropTypes.string,
-  /**
-  * Indicates if the icon spinner should be animated.
-  */
-  isAnimated: PropTypes.bool,
-  /**
-  * The visual theme to be applied to the overlay background. Accepts 'light', 'dark', and 'clear' or BackgroundStyles.LIGHT, BackgroundStyles.DARK, and BackgroundStyles.CLEAR.
-  */
+   * The visual theme to be applied to the overlay background. Accepts 'light', 'dark', and 'clear' or BackgroundStyles.LIGHT, BackgroundStyles.DARK, and BackgroundStyles.CLEAR.
+   */
   backgroundStyle: PropTypes.oneOf(['light', 'dark', 'clear', BackgroundStyles]),
   /**
-  * Indicates if the overlay is open.
-  */
+   * Indicates if the icon spinner should be animated.
+   */
+  isAnimated: PropTypes.bool,
+  /**
+   * Indicates if the overlay is open.
+   */
   isOpen: PropTypes.bool,
   /**
-  * Indicates if the overlay is relative to the triggering container.
-  */
+   * Indicates if the overlay is relative to the triggering container.
+   */
   isRelativeToContainer: PropTypes.bool,
+  /**
+   * The message to be displayed within the overlay.
+   */
+  message: PropTypes.string,
   /**
    * Used to select the root mount DOM node. This is used to help prevent focus from shifting outside of the overlay when it is opened in a portal.
    */
@@ -49,40 +49,34 @@ const defaultProps = {
   rootSelector: '#root',
 };
 
-const contextTypes = {
-  /* eslint-disable consistent-return */
-  intl: (context) => {
-    if (context.intl === undefined) {
-      return new Error('Component is internationalized, and must be wrapped in terra-base');
-    }
-  },
-  /* eslint-enable consistent-return */
-};
-
 const LoadingOverlay = ({
-  message,
   isAnimated,
+  message,
   rootSelector,
   ...customProps
-}, {
-  intl,
 }) => {
   // eslint-disable-next-line no-param-reassign
   delete customProps.onRequestClose;
 
-  const loadingMessage = message !== undefined ? message : intl.formatMessage({ id: 'Terra.Overlay.loading' });
-
   return (
     <Overlay {...customProps} className={cx('loading-overlay', customProps.className)} rootSelector={rootSelector}>
       <IconSpinner className={cx('icon')} isSpin={isAnimated} height="36" width="36" />
-      <div className={cx('message')}>{loadingMessage}</div>
+      {message !== undefined
+        ? <div className={cx('message')}>{message}</div>
+        : (
+          <FormattedMessage id="Terra.Overlay.loading">
+            {loadingMessage => (
+              <div className={cx('message')}>{loadingMessage}</div>
+            )}
+          </FormattedMessage>
+        )
+      }
     </Overlay>
   );
 };
 
 LoadingOverlay.propTypes = propTypes;
 LoadingOverlay.defaultProps = defaultProps;
-LoadingOverlay.contextTypes = contextTypes;
 LoadingOverlay.Opts = Overlay.Opts;
 
 export default LoadingOverlay;