+++ packages/terra-search-field/src/terra-dev-site/doc/example/SearchFieldDisabled.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
-import SearchFieldExampleTemplate from 'terra-search-field/lib/terra-dev-site/doc/example/SearchFieldExampleTemplate';
+import SearchField from 'terra-search-field/lib/SearchField';
 
 const SearchFieldDisabled = () => (
-  <SearchFieldExampleTemplate
-    isDisabled
-  />
+  <SearchField isDisabled />
 );
 
 export default SearchFieldDisabled;