+++ packages/terra-toggle/src/terra-dev-site/test/toggle/DefaultToggle.test.jsx
@@ -14,9 +14,10 @@ class ToggleDefault extends React.Component {
   }
 
   render() {
+    const { isOpen } = this.state;
     return (
       <div>
-        <button type="button" id="trigger-toggle" onClick={this.handleOnClick}>Default Toggle</button>
+        <button type="button" id="trigger-toggle" onClick={this.handleOnClick} aria-expanded={isOpen} aria-controls="toggle">Default Toggle</button>
         <Toggle id="toggle" isOpen={this.state.isOpen}>
           <p>
             Lorem ipsum dolor sit amet,