+++ packages/terra-dialog/CHANGELOG.md
@@ -3,7 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.22.0 - (August 14, 2019)
+------------------
 ### Changed
+* Components updated to use `FormattedMessage` to interface with `react-intl's` `intl` context.
 * updated package.json test scripts
 
 2.21.0 - (July 30, 2019)
@@ -34,6 +38,7 @@ Unreleased
 2.16.0  - (July 11, 2019)
 ------------------
 ### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports
 * Updated Docs around the requirement of terra-base and aggregate-translations.